only changed the name
This is the exact same file as the other one in the repository, but we did not name our file correctly, so this one is the exact same, but with a name change. I (Hao Chen) wrote you (Shu Hao) an e-mail about this issue. 
This is just in case correctors have scripts that need to find the file under a specific name, we have it, just under the wrong name, but this is the exact same file.